Move chunk decoding into anvil
Currently much non-drawing functionality is in the draw module
and not easily available to external code.

This code change results in chunks not being iterated in the
same order as they exist in the region file.
To my astonishment this makes `anvil render` run 10-20% faster
on my MacBook Pro with a World consisting of 30 Region Files.

It also tries to add at least some documentation.