Optionally use heightmap from chunk.
Using the heightmap stored in the actual chunks is much faster, so
we switch to this in the general case, making an option to manually
calculate the heightmap in the rare circumstances that you need to.

The only example of broken heightmaps I have come across is Etho's
lets play world.